Improve routes handling and logging, and add support for configuration file
This commit improves the handling of routes and logging in the `Route`
class. It introduces a new `config` method that allows the user to
specify configurations in a `phpslides.config.json` file. This method
disallows navigation to wrong paths or files for increased security.
Additionally, a new `$log` property is added, allowing the user to
control whether request logging is enabled or not.

The commit also enhances the `Route::any` method to support handling all
types of requests, and adds support for array routes and URL parameters.

These changes improve the flexibility and security of the routing system
and make it easier to configure and manage routes.

Resolves: #123